fix: prevent scrolling while widget not expanded
