+++ src/hooks/useWidgetExpanded.ts
@@ -1,0 +1,19 @@
+import { WidgetEvent, useWidgetEvents } from '@lifi/widget';
+import { useEffect, useState } from 'react';
+
+export const useWidgetExpanded = () => {
+  const [widgetExpanded, setWidgetExpanded] = useState(false);
+  const widgetEvents = useWidgetEvents();
+
+  useEffect(() => {
+    const handleWidgetExpanded = async (expanded: boolean) => {
+      setWidgetExpanded(expanded);
+    };
+    widgetEvents.on(WidgetEvent.WidgetExpanded, handleWidgetExpanded);
+
+    return () =>
+      widgetEvents.off(WidgetEvent.WidgetExpanded, handleWidgetExpanded);
+  }, [widgetEvents, widgetExpanded]);
+
+  return widgetExpanded;
+};
